avoid some PWM syncing issues, but unfortunately does not fix #578

+++ src/windows/window.c
@@ -413,7 +413,7 @@ int win_match_refresh;
 int win_sync_refresh;
 float win_gfx_brightness;
 int win_blit_effect;
-float win_screen_aspect = (4.0f / 3.0f);
+float win_screen_aspect = (float)(4.0 / 3.0);
 
 // windows
 HWND win_video_window;
